refactor: extract validation and migration utilities
- Extract type guard utilities to utils/validation.ts for better maintainability
- Move migration logic to utils/homepage-state-migration.ts for improved readability
- Simplify homepage-state-service.ts by using dedicated utility modules
- Addresses Copilot review feedback for better code organization
