+++ src/services/homepage-state-service.ts
@@ -1,49 +1,13 @@
 import { UnifiedDomainResult } from '@/types/domain';
+import { isValidNumber } from '@/utils/validation';
+import {
+  migrateHomepageState,
+  SCHEMA_VERSION,
+} from '@/utils/homepage-state-migration';
 
 // Constants for localStorage keys
 const HOMEPAGE_STATE_KEY = 'domain-hunt-homepage-state';
 const STATE_EXPIRY_HOURS = 24; // State expires after 24 hours
-
-// Current schema version for migration tracking
-const SCHEMA_VERSION = 1;
-
-// Default collapsed categories (all except popular)
-const DEFAULT_COLLAPSED_CATEGORIES = [
-  'international',
-  'academic---education',
-  'finance',
-  'professional',
-  'businesses',
-  'audio---video',
-  'arts---culture',
-  'adult',
-  'marketing',
-  'technology',
-  'real-estate',
-  'short',
-  'politics',
-  'organizations',
-  'shopping-and-sales',
-  'media-and-music',
-  'fun',
-  'sports-and-hobbies',
-  'products',
-  'transport',
-  'personal',
-  'social---lifestyle',
-  '-2-or-less',
-  'food---drink',
-  'services',
-  'beauty',
-  'cities',
-  'travel',
-  'health-and-fitness',
-  'colors',
-  'trades---construction',
-  'non-english',
-  'religion',
-  'miscellaneous',
-];
 
 // Types for persisted state
 export interface HomepageState {
@@ -90,40 +54,6 @@ export const saveHomepageState = (
   }
 };
 
-// Validation helpers
-const isValidStringArray = (value: unknown): value is string[] =>
-  Array.isArray(value) && value.every(item => typeof item === 'string');
-const isValidBoolean = (value: unknown): value is boolean =>
-  typeof value === 'boolean';
-const isValidNumber = (value: unknown): value is number =>
-  typeof value === 'number' && !isNaN(value);
-const isValidString = (value: unknown): value is string =>
-  typeof value === 'string';
-
-// Migration function for old format to new format
-const migrateFromV0ToV1 = (
-  rawState: Record<string, unknown>
-): Partial<HomepageState> => {
-  // V0 format: { domains, selectedTlds, unifiedResult, savedAt }
-  // V1 format: adds { collapsedCategories, showAllCategories, searchQuery, version }
-
-  return {
-    // Preserve existing fields exactly as they are
-    domains: isValidStringArray(rawState.domains) ? rawState.domains : [],
-    selectedTlds: isValidStringArray(rawState.selectedTlds)
-      ? rawState.selectedTlds
-      : [],
-    unifiedResult: rawState.unifiedResult as UnifiedDomainResult | null,
-    savedAt: isValidNumber(rawState.savedAt) ? rawState.savedAt : Date.now(),
-
-    // Set intelligent defaults for new fields
-    collapsedCategories: DEFAULT_COLLAPSED_CATEGORIES,
-    showAllCategories: false,
-    searchQuery: undefined,
-    version: SCHEMA_VERSION,
-  };
-};
-
 export const loadHomepageState = (): Partial<HomepageState> | null => {
   if (typeof window === 'undefined') return null;
 
@@ -159,38 +89,10 @@ export const loadHomepageState = (): Partial<HomepageState> | null => {
     }
 
     // Determine version and migrate if necessary
-    const version = rawState.version || 0; // Default to 0 for old format
-    let migratedState: Partial<HomepageState>;
+    const version = isValidNumber(rawState.version) ? rawState.version : 0; // Default to 0 for old format
+    const migratedState = migrateHomepageState(rawState, version);
 
-    if (version === 0) {
-      // Migrate from V0 (old format) to V1 (current format)
-      migratedState = migrateFromV0ToV1(rawState);
-      console.log('Migrated homepage state from V0 to V1');
-    } else if (version === SCHEMA_VERSION) {
-      // Current version, validate and use as-is
-      migratedState = {
-        domains: isValidStringArray(rawState.domains) ? rawState.domains : [],
-        selectedTlds: isValidStringArray(rawState.selectedTlds)
-          ? rawState.selectedTlds
-          : [],
-        unifiedResult: rawState.unifiedResult as UnifiedDomainResult | null,
-        savedAt: isValidNumber(rawState.savedAt)
-          ? rawState.savedAt
-          : Date.now(),
-        collapsedCategories: isValidStringArray(rawState.collapsedCategories)
-          ? rawState.collapsedCategories
-          : DEFAULT_COLLAPSED_CATEGORIES,
-        showAllCategories: isValidBoolean(rawState.showAllCategories)
-          ? rawState.showAllCategories
-          : false,
-        searchQuery: isValidString(rawState.searchQuery)
-          ? rawState.searchQuery
-          : undefined,
-        version: SCHEMA_VERSION,
-      };
-    } else {
-      // Unknown version, treat as corrupted
-      console.warn(`Unknown homepage state version: ${version}, removing data`);
+    if (!migratedState) {
       localStorage.removeItem(HOMEPAGE_STATE_KEY);
       return null;
     }
